pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-audio-classification"
version = "0.1.0"
description = "Kubernetes-deployable ML audio classification experiment application"
authors = [
    {name = "DSE Soundhub Team", email = ""}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0,<1.25.0",
    "pandas>=2.0.0,<2.1.0",
    "scikit-learn>=1.6.1,<2.0.0",
    "scipy>=1.10.0,<1.12.0",
    "matplotlib>=3.7.0,<3.8.0",
    "seaborn>=0.12.0",
    "pydantic>=2.3.0,<2.4.0",
    "pydantic-settings>=2.0.0,<2.1.0",
    "google-cloud-storage>=2.10.0,<2.11.0",
    "google-cloud-logging>=3.6.0,<3.7.0",
    "tensorflow>=2.13.0,<2.15.0",
    "tensorflow-hub>=0.15.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "timm>=0.9.0",
    "librosa>=0.10.0,<0.11.0",
    "soundfile>=0.12.0,<0.13.0",
    "opensoundscape>=0.10.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "aiofiles>=23.0.0,<24.0.0",
    "click>=8.1.0,<8.2.0",
    "python-dotenv>=1.0.0,<1.1.0",
    "structlog>=23.1.0,<23.2.0",
    "rich>=13.5.0,<13.6.0",
    "typer>=0.9.0,<0.10.0",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
]

[project.scripts]
ml-audio-classify = "ml_audio_classification.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=ml_audio_classification --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]